Exercise 4.21

+++ part04/controllers/blogs.js
@@ -37,6 +37,22 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+
+  if (!blog) {
+    return response.status(401).json({ error: 'blog not found' })
+  }
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!decodedToken || !decodedToken.id || !blog.user) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  if (decodedToken.id.toString() != blog.user.toString()) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id);
   response.status(204).end();
 })
